requirements.txt
numpy~=2.2.3
matplotlib~=3.10.1
opencv-python~=4.11.0.86
pillow~=12.0.0
torch~=2.6.0
torchvision~=0.19.0
tqdm~=4.66.0
tensorboard~=2.20.0